Initial take of the LoaderApp.
This is an experiment in splitting apps
into components that are finer grained
than Activities so that the components
can be mixed and matched on large screens.

The code is lossely based on the Contacts
application, reusing much of the logic
for displaying contact details and none
of the logic for the contacts list.

Change-Id: I74f1c438184cf260e03fa3f2fe7a2fee395d9dbc