LoaderActivity now tracks Loaders with the same ID.
If you start a load with the same ID as a previously
existing loader the old loader will now be destroyed
after the new loader completes it's work instead
of when the new laoder is started. This prevents
flickering in UIs that draw the empty state
when their data is missing, like ListView.

Change-Id: Iee7ae3c57521ec1f2de246b9a8f72f25dd931c86