Enhance groups mode and add a view contact intent handler.
Groups now uses a loader and coupler and tries
to load the full list of existing groups to
choose from as well as All Contacts and Favs.

The ViewContact activity catches the standard
platform intent to view a contact and redirects
to the proper UI based on the screen class.
This will hopefully be done differently, but
shows one possible way to handle different
layouts.

Various layout cleanups.

Moved ListCoupler to the patterns pacakge as
it mave prove to be useful across apps.

Change-Id: I2392860c20dcbf73bf648479225ca9ac917b57c9