UI automation service
A UI automation service is included, it uses accessibility service
to perform UI verification, and uses WindowManager (with special
permission) to perform event injection. This will only work on
userdebug builds.

A library is also provided, which attempts to hide the details of
starting and stopping the service, also provides easy access to
the service via dependency injection by providing a customized test
runner. The library also provides ability to trace all calls to
the service and store them for later investigation on failures.

Change-Id: Iff8c4070dfd5c3bca30b1026600d654391c90d34