Script for SELinux CTS avc rules test.
Takes an input SELinux policy.conf file and generates
an XML file based on the allow and neverallow rules.  The file contains rules,
which are created by expanding the SELinux rule notation into the individual
components which a checkAccess() check, that a policy manager would have to
perform, needs.

It may be replaced in the future by relying on checkpolicy as a front end, by
potentially using the compiled policy as input instead, or by expanding it
into a much more general parser, which would be more easily extended to new
purposes.

Change-Id: Ifb1a8a2d492d8ffc211352afed7daa7bd80b07b1