Multi packages per APK sample
This is a sample that shows how to declare multi packages per APK.
The feature is currently supported only for privileged apps or updates
to such apps. The sample declares three activities wihch on old
platforms are in the parent package but on devices running API 24 two
of the activities are disabled in the parent and moved to the first
and second child packages respectively.

Change-Id: I673d9c455b3c8543091e83e8d37e35c4af1067d7